Move parse_path
