+++ skyplane/cli/cli_impl/cp_replicate_fallback.py
@@ -1,7 +1,7 @@
 import os
 from typing import Optional
 
-from skyplane.cli.common import parse_path
+from skyplane.api.impl.path import parse_path
 
 
 def fallback_cmd_local_cp(src_path: str, dest_path: str, recursive: bool) -> str:
